Throw exception if request for provider courses fails
The AttributeError exception that we were seeing before was due to the
Slumber API constructor never being called. There was a variable within
the class that was referenced elsewhere in code, _store, that was
unavailable without this.

I uncommented the constructor, and removed the check to see if the API
is specifically the "Other" API client. Instead, the exception is caught
and displayed in the adaptive bridge.